// mobile/src/api/generated/model/limit.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Firetower
 * The Firetower control plane: API, scheduling, and worker transports.
 * OpenAPI spec version: 0.38.1
 */

export interface Limit {
  /** @nullable */
  resetsAt?: number | null;
  scope: string;
  status: string;
  /** @nullable */
  usedPercent?: number | null;
}