// mobile/src/api/generated/model/sessionAccount.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Firetower
 * The Firetower control plane: API, scheduling, and worker transports.
 * OpenAPI spec version: 0.36.2
 */
import type { Account } from './account';
import type { Limit } from './limit';
import type { Switch } from './switch';

export interface SessionAccount {
  account?: null | Account;
  limits: Limit[];
  switches: Switch[];
}